backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sse-starlette>=2.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Authentication
authlib>=1.3.0
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Research Engine
requests>=2.31.0
beautifulsoup4>=4.12.3
camoufox[geoip]>=0.4.0
ddgs>=6.0.0

# Utils
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
aiofiles>=23.2.1

# Export
jinja2>=3.1.3
weasyprint>=61.0  # PDF generation
markdown>=3.5.2
